// src/app/pages/transfer/drivers-transfer/list-drivers/shared/list-drivers-enum.ts
export enum ListDriversEnum {        
    driverNameTitle =  'Driver Name',
    licenceIdTitle = 'Driver Licence ID',
    contactPhoneTitle = 'Phone',
    birthdayTitle = 'Birthdate',
    actionTitle = 'Action',

    driverName =  'driverName',
    driverLicenceId = 'driverLicenceId',
    contactPhone = 'contactPhone',
    birthday = 'birthday',
    action = 'action',

    model = 'model',

    valueZero = 0,
    valueOne = 1,
    valueTwo = 2,
}